requirements.txt
pandas
groq
python-dotenv
openpyxl
boto3